Improve compaction efficiency.
Databases that are updated randomly by doc id cause compaction to be
noticably less efficient than possible. This is due to the order in
which the id_tree is written. The random updates cause the append only
btree to generate a noticeable amount of waste.

This patch works by writing the id_tree to a temporary file during
compaction and then streams this btree back to the compaction file
during a final pass.