Implementation of efficient btree copying.
If we have a source of sorted k/v data its possible to efficiently
construct a btree by streaming them to disk and building the upper btree
nodes on demand. This process also has the nice property that it results
in zero garbage for the new btree.

The two downsides to this process are that it can't be resumed if
interrupted and can not be merged into an existing btree.

The couch_btree build API looks like this:

    % Some preparation
    {ok, Fd} = couch_file:open("newbtree.dat", [create]),
    {ok, InitBtree} = couch_btree:open(nil, Fd),

    % Build the new tree.
    Init = couch_btree:build_init(InitBtree),
    Final = lists:foldl(fun(I, Acc) ->
        couch_btree:build_add({I, I}, Acc)
    end, Init, lists:seq(1, 1000)),
    {ok, NewBtree} = couch_btree:build_finish(Final).

Original concept from Filipe Manana