Don't always require exactly N replies for an mp attachment PUT
It's not safe to assume we require, or will receive, exactly N replies
(where N is read from the "n" key of the "cluster" section of the
configuaration). This needs proper fabric-ification.

This commit will at least allow replication tests with clusters of
less than N nodes where the documents have attachments (which triggers
the multipart code).

BugzID: 14258